feat: moved alg to module
