+++ src/encoding.h
@@ -1,0 +1,7 @@
+#ifndef ENCODING_H
+#define ENCODING_H
+
+// Function to encrypt the plaintext using the Vigenère cipher algorithm
+char *vigenereEncrypt(const char plaintext[], const char keyword[]);
+
+#endif /* ENCODING_H */
